refactor(player): move media element error circuit breaker to internal/helpers
The circuit breaker is a small reusable utility (similar in character
to retry.ts), engine-agnostic at its core, consumed by both ShakaPlayer
and BrowserPlayer. internal/helpers/ is a better fit than co-locating
it next to one player engine in src/player/.

Pure move — no behaviour change. Only the relative imports inside the
helper and the import paths in shakaPlayer.ts / browserPlayer.ts had
to be updated.
